--- noFreeRider/src/main/java/com/teamnine/noFreeRider/Member/domain/MemberTaskRepository.java
@@ -1,7 +1,0 @@
-package com.teamnine.noFreeRider.Member.domain;
-
-import com.teamnine.noFreeRider.Member.repository.MemberTask;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberTaskRepository extends JpaRepository<MemberTask, Long> {
-}
