--- noFreeRider/src/main/java/com/teamnine/noFreeRider/Member/repository/MemberProjectRepository.java
@@ -1,7 +1,0 @@
-package com.teamnine.noFreeRider.Member.repository;
-
-import com.teamnine.noFreeRider.Member.domain.MemberProject;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberProjectRepository extends JpaRepository<MemberProject,Long> {
-}
