--- noFreeRider/src/main/java/com/teamnine/noFreeRider/Member/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.teamnine.noFreeRider.Member.repository;
-
-import com.teamnine.noFreeRider.Member.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface MemberRepository extends JpaRepository<Member, UUID> {
-}
