--- noFreeRider/src/main/java/com/teamnine/noFreeRider/Task/repository/TaskRepository.java
@@ -1,9 +1,0 @@
-package com.teamnine.noFreeRider.Task.repository;
-
-import com.teamnine.noFreeRider.Task.domain.Task;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface TaskRepository extends JpaRepository<Task, UUID> {
-}
