--- noFreeRider/src/main/java/com/teamnine/noFreeRider/Web/controller/authenticationController.java
@@ -1,4 +1,0 @@
-package com.teamnine.noFreeRider.Web.controller;
-
-public class authenticationController {
-}
