+++ noFreeRider/src/main/java/com/teamnine/noFreeRider/member/domain/MemberProject.java
@@ -1,7 +1,6 @@
-package com.teamnine.noFreeRider.Member.domain;
+package com.teamnine.noFreeRider.member.domain;
 
-import com.teamnine.noFreeRider.Member.domain.Member;
-import com.teamnine.noFreeRider.Project.domain.Project;
+import com.teamnine.noFreeRider.project.domain.Project;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
