+++ noFreeRider/src/main/java/com/teamnine/noFreeRider/member/repository/MemberProjectRepository.java
@@ -1,0 +1,9 @@
+package com.teamnine.noFreeRider.member.repository;
+
+import com.teamnine.noFreeRider.member.domain.MemberProject;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberProjectRepository extends JpaRepository<MemberProject,Long> {
+}
