+++ noFreeRider/src/main/java/com/teamnine/noFreeRider/member/repository/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.teamnine.noFreeRider.member.repository;
+
+import com.teamnine.noFreeRider.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, UUID> {
+}
