+++ noFreeRider/src/main/java/com/teamnine/noFreeRider/notification/repository/NotificationRepository.java
@@ -1,0 +1,11 @@
+package com.teamnine.noFreeRider.notification.repository;
+
+import com.teamnine.noFreeRider.notification.domain.Notification;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface NotificationRepository extends JpaRepository<Notification, UUID> {
+}
