+++ noFreeRider/src/main/java/com/teamnine/noFreeRider/project/domain/Project.java
@@ -1,6 +1,6 @@
-package com.teamnine.noFreeRider.Project.domain;
+package com.teamnine.noFreeRider.project.domain;
 
-import com.teamnine.noFreeRider.Member.domain.Member;
+import com.teamnine.noFreeRider.member.domain.Member;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
