+++ noFreeRider/src/main/java/com/teamnine/noFreeRider/project/repository/ProjectRepository.java
@@ -1,0 +1,11 @@
+package com.teamnine.noFreeRider.project.repository;
+
+import com.teamnine.noFreeRider.project.domain.Project;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface ProjectRepository extends JpaRepository<Project, UUID> {
+}
