+++ noFreeRider/src/main/java/com/teamnine/noFreeRider/task/repository/TaskRepository.java
@@ -1,0 +1,11 @@
+package com.teamnine.noFreeRider.task.repository;
+
+import com.teamnine.noFreeRider.task.domain.Task;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface TaskRepository extends JpaRepository<Task, UUID> {
+}
