+++ noFreeRider/src/main/java/com/teamnine/noFreeRider/web/controller/AuthenticationController.java
@@ -1,0 +1,4 @@
+package com.teamnine.noFreeRider.web.controller;
+
+public class AuthenticationController {
+}
