+++ noFreeRider/src/main/java/com/teamnine/noFreeRider/web/controller/PageController.java
@@ -1,10 +1,10 @@
-package com.teamnine.noFreeRider.Web.controller;
+package com.teamnine.noFreeRider.web.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class pageController {
+public class PageController {
     @GetMapping("/login")
     public String login() {
         return "login";
